% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/batchOutFiles.R
\name{batchOutFiles}
\alias{batchOutFiles}
\title{Process Batch Out Files For Sortie}
\usage{
batchOutFiles(namepattern = "", sdir = ".", yearoffset = 0)
}
\arguments{
\item{namepattern}{A character string that matches the unique part of your
batch files. E.g., if all files in a particular batch had "monkey" in them,
this function would search any file in that directory with the string
"monkey" inside of it.}

\item{sdir}{The path to the directory to search. Default is ".", representing
your current working directory.}

\item{yearoffset}{The amount to subtract from the "step" variable to account
for intial plot loading.}
}
\value{
This function returns a data.frame of cleaned and merged files. Any
  personally identifying information will be lost (e.g., filename).
}
\description{
This is a function to process batches of .out files generated by SORTIE-ND's
summary output. It loops through a list of filenames given and produces a
single data.frame as output. This function relies on \code{\link{getOutFile}}
and \code{\link{parseOutFile}}.
}

