% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/changeColsBySpp.R
\name{changeColsBySpp}
\alias{changeColsBySpp}
\title{Extract Column Info And Rearrage A Data.Frame}
\usage{
changeColsBySpp(df, removalkey, dead = F)
}
\arguments{
\item{df}{The data.frame to search}

\item{removalkey}{The repeated part of the column name to remove, e.g.,
"Density"}

\item{dead}{If dead=T, append "Natural.Dead." to the removal key and subset
that way.}
}
\description{
This function takes columns that are names like (DensityABCO, DensityABMA,
DensityCADE) and changes the data.frame into something like (Species,
Density), with a key/value system instead. Check out the examples for how it
works. It's used within \code{\link{parseOutFile}}.
}
\examples{
testdf <- data.frame(DensityABCO=c(1,2), DensityABMA=c(3,4), DensityCADE=c(5,6), stringsAsFactors=F)
changeColsBySpp(testdf, "Density", dead=F)
}

