% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/findPlotStart.R
\name{findPlotStart}
\alias{findPlotStart}
\title{Find The Starting Place For A Map Centered In A Larger Map}
\usage{
findPlotStart(newmapx, newmapy, plotx, ploty)
}
\arguments{
\item{newmapx}{The range of x in your larger, new map.}

\item{newmapy}{The range of y in your larger, new map.}

\item{plotx}{The range of x in your smaller map.}

\item{ploty}{The range of y in your smaller map.}
}
\description{
This function finds the place where a smaller range of (x,y) should begin in
order to be centered in a larger (x,y) space.
}
\examples{
findPlotStart(300, 300, 100, 150)
}

