% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/getOutFile.R
\name{getOutFile}
\alias{getOutFile}
\title{Import a SORTIE-ND .out File Into R}
\usage{
getOutFile(filenames, numsubplots = 1)
}
\arguments{
\item{filenames}{A vector or character string of the filenames and paths you
want to import. Include path if you're not in directory where the files are
located.}
}
\value{
This function returns a single data.frame of all the values. If your
  input files are not the same column number, this will throw an error.
}
\description{
This function removes the first 5 lines of an .out file and then imports it
into a data.frame for further work.
}
\examples{
# getOutFile("myoutfile.out")
 # getOutFile(c("myoutfile1.out", "myoutfile2.out"))
}

