% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/getRange.R
\name{getRange}
\alias{getRange}
\title{Get the difference between two numbers}
\usage{
getRange(x)
}
\value{
This function returns a number.
}
\description{
A simple function to calculate the difference between the maximum value and the minimum value of x.
}
\examples{
getRange(c(12, 2,4))
}

