% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/parseOutFile.R
\name{parseOutFile}
\alias{parseOutFile}
\title{Parse a data.frame imported from a SORTIE-ND OUT File.}
\usage{
parseOutFile(df, removeSpp = c("NR", "Total"))
}
\arguments{
\item{df}{The data.frame generated by \code{\link{getOutFile}}.}

\item{removeSpp}{Optional, already set to remove "NR" and "Total" columns
from reformatting, so you can deal with those separately or regenerate them
yourself.}
}
\description{
This function takes the data.frame generated by \code{\link{getOutFile}} and
reformats it so that Step, Subplot, and Species are the first 3 columns,
followed by values for Absolute Densities and Basal Areas. By default, this
function removes any columns labeled "NR" or "Total".
}
\note{
This function cannot handle the relative density or basal areas
  statistics provided by SORTIE-ND, possibly because "rel" or some form of it
  is a protected word for the \code{\link{grep}} family of functions that I
  use.
}
\examples{
#dat <- getOutFile("test.out")
#head(parseOutFile(dat))
}

