% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plotExpSim.R
\name{plotExpSim}
\alias{plotExpSim}
\title{Plot Expected And Simulated Values}
\usage{
plotExpSim(expDf, simDf, charactername, means = F, filename = "",
  exec = F, exportSimDf = F)
}
\arguments{
\item{expDf}{This is a data.frame of expected results for a plot. Must contain a "Step" column (x-axis), a "Species" column (color factor), and a value column that has the name of the "charactername" parameter.}

\item{simDf}{This is a data.frame of simulated results for a plot. Must contain a "Step" column (x-axis), a "Species" column (color factor), and a value column that has the name of the "charactername" parameter.}

\item{charactername}{The name of the column to plot on the y-axis.}

\item{means}{Currently set to false (default). If TRUE, function will plot means of simulated values instead of all values in simDf.}

\item{filename}{The filename to write this plot to. Include path if necessary.}

\item{exec}{If exec is TRUE, then the program will write the plot image to file.}

\item{exportSimDf}{If exportSimDf is TRUE, then the function will return a data.frame of the simDf. This is most often used when means=T, and the function returns the means of simDf.}
}
\value{
This function
}
\description{
This function will take two data.frames, a simulated and an expected data.frame, and plots the points over top of each other to visualize just how closely they match.
}

