% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/processTestPlot.R
\name{processTestPlot}
\alias{processTestPlot}
\title{Process a plot's test data}
\usage{
processTestPlot(plotname, sdir, yearoffset, charactername = "",
  writeimage = NULL, writefile = NULL, byspecies = F, numsubplots = 1,
  subplotid = 0)
}
\description{
Process a plot's test data
}

