% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/reNumber.R
\name{reNumber}
\alias{reNumber}
\title{Renumber rows in a data.frame to deal with missing rows.}
\usage{
reNumber(x)
}
\arguments{
\item{x}{The target data.frame}
}
\value{
This function returns a data.frame
}
\description{
This function cleans up the rownames of a data.frame by renumbering from 1 to the total number of rows in the data.frame.
}
\examples{
reNumber(data.frame(x=1:4, y=5:8)[-1,])
}

