% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/rotPlots.R
\docType{data}
\name{rotPlots}
\alias{rotPlots}
\title{Rotated Plots}
\format{\preformatted{'data.frame':	51216 obs. of  13 variables:
 $ X        : int  2 3 4 5 6 7 8 9 10 11 ...
 $ plot     : chr  "crackers" "crackers" "crackers" "crackers" ...
 $ subplot  : int  1 1 1 1 1 1 1 1 1 1 ...
 $ treeid   : int  1 1 1 1 1 2 2 2 2 2 ...
 $ species  : chr  "ABCO" "ABCO" "ABCO" "ABCO" ...
 $ ingrowth : int  NA NA NA NA NA NA NA NA NA NA ...
 $ firstrec : int  1992 1992 1992 1992 1992 1992 1992 1992 1992 1992 ...
 $ deathyear: int  NA NA NA NA NA NA NA NA NA NA ...
 $ x        : num  19 19 19 19 19 23 23 23 23 23 ...
 $ y        : num  81 81 81 81 81 76 76 76 76 76 ...
 $ measyear : int  1992 1997 2002 2007 2012 1992 1997 2002 2007 2012 ...
 $ dbh      : num  4.4 5 5.6 6 6.2 8.7 9.1 9.8 10.4 11.1 ...
 $ stage    : chr  "tree" "tree" "tree" "tree" ...
}}
\usage{
rotPlots
}
\description{
This data.frame is a subset of expandedTrees from MakeMyForests, rotated with
functions found in the disperseR package. It is only 11 plots. See
expandedTrees in MakeMyForests for a detailed description of each column.
}
\keyword{datasets}

