% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/shiftPlot.R
\name{shiftPlot}
\alias{shiftPlot}
\title{Shift Plot Linearly By X And Y}
\usage{
shiftPlot(oldx, oldy, newx, newy, x, y)
}
\arguments{
\item{oldx}{Your old origin's x point (e.g., 0)}

\item{oldy}{Your old origin's y point}

\item{newx}{Your new origin x.}

\item{newy}{Your new oriign y.}

\item{x}{A vector of x-values to be transformed.}

\item{y}{A vector of y-values to be transformed.}
}
\description{
This function takes your old origin, your new origin, and your data.frame, and shifts to match the new origin.
}
\examples{
shiftPlot(0,0,100,100,c(2,4,5), c(4,5,6))
}

