% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/sortieOutputs.R
\docType{Package}
\name{sortieOutputs}
\alias{sortieOutputs}
\title{sortieOutputs: A package for parsing SORTIE-ND output files}
\description{
This package provides parsing support for SORTIE-ND summary output files.
These output files are essentially giagantic tab-delimited files that are
very hard for a naive computer program to read. This package will take a
folder of output files and convert them into a more-readable CSV file which
can then be imported into R or other statistics programs for analysis.
}

