% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/ssdSeedlings.R
\docType{data}
\name{ssdSeedlings}
\alias{ssdSeedlings}
\title{ssdSeedlings}
\format{\preformatted{'data.frame':	36773 obs. of  6 variables:
 $ plot      : chr  "bellow" "bellow" "bellow" "bellow" ...
 $ subplot   : int  8 8 8 8 8 8 8 8 8 8 ...
 $ sppcode   : chr  "PICO" "PICO" "PICO" "PICO" ...
 $ maxht     : int  10 10 10 10 10 10 10 10 10 10 ...
 $ year      : chr  "1999" "2000" "2001" "2002" ...
 $ nSeedlings: int  0 0 2 0 0 0 0 0 0 0 ...
}}
\usage{
ssdSeedlings
}
\description{
A data.frame of seedling counts for 10 years (1999-2009) in the USGS plots.
}
\keyword{datasets}

