% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/testExpSim.R
\name{testExpSim}
\alias{testExpSim}
\title{Test For Differences Between Actual and Simulated Data}
\usage{
testExpSim(expDf, simDf, charactername, mymu = NULL, write = F,
  filename = "")
}
\arguments{
\item{expDf}{An experimental data.frame with columns "Step", "Species", and a
column named as the value inputted in "charactername" parameter.}

\item{simDf}{A data.frame with simulated values, with columns "Step",
"Species", and a column named as the value inputted in "charactername"
parameter.}

\item{charactername}{The name of the column of interest, e.g.,
"AdultAbsoluteDensity"}

\item{mymu}{If mymu is set to something other than null, it will be used as
the mu value for a one-sample t-test. If left blank, means from expDf will
be used.}

\item{write}{If write is set to true, this function will write a CSV file
with filename provided}

\item{filename}{If write is set to true, this function will write a CSV file
  with this filename.}
}
\value{
This function returns a data.frame with Step, Species, charactername,
  pval, and signif columns.
}
\description{
This function will perform one-sample t-tests, with a manually set mu, for
the data that you enter. Even if you want to set a manual mu value, you need
to include an experimental / expected Df.
}

