% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/testOutImport.R
\docType{data}
\name{testOutImport}
\alias{testOutImport}
\title{testOutImport}
\format{\preformatted{'data.frame':	18 obs. of  8 variables:
 $ Step       : num  1 1 1 1 1 1 1 1 1 1 ...
 $ Subplot    : int  0 1 0 1 0 1 0 1 0 1 ...
 $ Species    : Factor w/ 9 levels "ABCO","ABMA",..: 1 1 3 3 5 5 7 7 8 8 ...
 $ SdlAbsDen  : num  17.778 17.778 41.222 41.222 0.111 ...
 $ SaplAbsDen : num  109.2 109.2 378.9 378.9 57.8 ...
 $ AdultAbsDen: num  45.44 45.44 309 309 3.56 ...
 $ SaplAbsBA  : num  1.227 1.227 3.029 3.029 0.976 ...
 $ AdultAbsBA : num  7.915 7.915 22.415 22.415 0.496 ...
}}
\usage{
testOutImport
}
\description{
A dataset for testing output from the batch OutFile function. Not real data.
}
\keyword{datasets}

