% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/TryDTest.R
\name{TryDTest}
\alias{TryDTest}
\title{Tries a Dixon test for outliers.}
\usage{
TryDTest(...)
}
\arguments{
\item{...}{This function takes the parameters of a dixon test.}
}
\description{
This function returns a 1 if there's an error in the Dixon test, otherwise,
it reports a p-value.
}

