% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/TryTTest.R
\name{TryTTest}
\alias{TryTTest}
\title{Tries a t-test, returns 1 if there is an error}
\usage{
TryTTest(...)
}
\description{
Function taken from https://stat.ethz.ch/pipermail/r-help/2008-February/154167.html
}

